pyproject.toml
[project]
name = "tfbench"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.49.0",
    "dacite>=1.8.1",
    "fire==0.5.0",
    "funcy==2.0",
    "funcy-chain==0.2.0",
    "google-genai>=1.11.0",
    "groq==0.8.0",
    "hypothesis>=6.98.6",
    "markdown-to-json==2.1.2",
    "matplotlib>=3.8.3",
    "numpy>=1.26.4",
    "ollama>=0.2.1",
    "openai==1.75.0",
    "pathos>=0.3.3",
    "pylint>=3.3.6",
    "pytest>=8.0.0",
    "python-dotenv==1.0.1",
    "requests==2.32.3",
    "returns[compatible-mypy]==0.22.0",
    "seaborn==0.13.2",
    "tabulate>=0.9.0",
    "tiktoken==0.7.0",
    "tqdm>=4.66.2",
    "tree-sitter==0.22.3",
    "tree-sitter-haskell==0.21.0",
    "types-requests>=2.31.0",
]

[build-system]
requires = ["uv_build>=0.8.10,<0.9.0"]
build-backend = "uv_build"
